use stricter errors

+++ src/core/horizon/display/driver.cpp
@@ -51,21 +51,21 @@ void Driver::Present(
 
     // Viewport
     const auto src_size = float2(OS_INSTANCE.GetDisplayResolution());
-    auto scale_x = f32(width) / src_size.x();
-    auto scale_y = f32(height) / src_size.y();
+    auto scale_x = static_cast<f32>(width) / src_size.x();
+    auto scale_y = static_cast<f32>(height) / src_size.y();
 
     FloatRect2D dst_rect;
     float dst_scale;
     if (scale_x > scale_y) {
         dst_scale = scale_y;
         const auto dst_width = src_size.x() * dst_scale;
-        dst_rect.origin = {(width - dst_width) / 2.f, 0.f};
-        dst_rect.size = {dst_width, f32(height)};
+        dst_rect.origin = {(static_cast<f32>(width) - dst_width) / 2.f, 0.f};
+        dst_rect.size = {dst_width, static_cast<f32>(height)};
     } else {
         dst_scale = scale_x;
         const auto dst_height = src_size.y() * dst_scale;
-        dst_rect.origin = {0.f, (height - dst_height) / 2.f};
-        dst_rect.size = {f32(width), dst_height};
+        dst_rect.origin = {0.f, (static_cast<f32>(height) - dst_height) / 2.f};
+        dst_rect.size = {static_cast<f32>(width), dst_height};
     }
 
     // Present
